Aplicando as regras de acesso

+++ app/Http/Controllers/Admin/UsuarioController.php
@@ -35,20 +35,25 @@ class UsuarioController extends Controller
     }
     public function index()
     {
-      /*if(auth()->user()->can('listar-usuarios',false)){
-        dd('sim');
+      if(auth()->user()->can('usuario_listar')){
+        $usuarios = User::all();
+        return view('admin.usuarios.index', compact('usuarios'));
       }else{
-        dd('nao');
-      }*/
-      $usuarios = User::all();
-      return view('admin.usuarios.index', compact('usuarios'));
+        return redirect()->route('admin.principal');
+      }
     }
     public function adicionar()
     {
+      if(!auth()->user()->can('usuario_adicionar')){
+        return redirect()->route('admin.principal');
+      }
       return view('admin.usuarios.adicionar');
     }
     public function salvar(Request $request)
     {
+      if(!auth()->user()->can('usuario_adicionar')){
+        return redirect()->route('admin.principal');
+      }
       $dados = $request->all();
       $usuario = new User();
       $usuario->name = $dados['name'];
@@ -64,11 +69,17 @@ class UsuarioController extends Controller
     }
     public function editar($id)
     {
+      if(!auth()->user()->can('usuario_editar')){
+        return redirect()->route('admin.principal');
+      }
       $usuario = User::find($id);
       return view('admin.usuarios.editar', compact('usuario'));
     }
     public function atualizar(Request $request,$id)
     {
+      if(!auth()->user()->can('usuario_editar')){
+        return redirect()->route('admin.principal');
+      }
       $usuario = User::find($id);
       $dados = $request->all();
       if(isset($dados['password']) && strlen($dados['password']) > 5)
@@ -87,6 +98,9 @@ class UsuarioController extends Controller
 
     public function deletar($id)
     {
+      if(!auth()->user()->can('usuario_deletar')){
+        return redirect()->route('admin.principal');
+      }
       User::find($id)->delete();
       \Session::flash('mensagem',[
         'msg'=> 'Registro deletado com sucesso',
